+++ axonius_api_client/__init__.py
@@ -3,7 +3,7 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 from . import api, auth, cli, constants, exceptions, http, logs, tools, version
-from .api import Adapters, Devices, Enforcements, Users
+from .api import Adapters, Devices, Enforcements, Users, System
 from .auth import ApiKey
 from .connect import Connect
 from .http import Http
@@ -23,6 +23,7 @@ __all__ = (
     "Devices",
     "Adapters",
     "Enforcements",
+    "System",
     # modules
     "api",
     "auth",