+++ axonius_api_client/api/__init__.py
@@ -2,19 +2,29 @@
 """Axonius API Client package."""
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-from . import adapters, enforcements, mixins, routers, users_devices
+from . import (
+    adapters,
+    enforcements,
+    mixins,
+    routers,
+    assets,
+    system,
+)
 from .adapters import Adapters
 from .enforcements import Enforcements
-from .users_devices import Devices, Users
+from .assets import Devices, Users
+from .system import System
 
 __all__ = (
     "Users",
     "Devices",
     "Adapters",
     "Enforcements",
+    "System",
     "routers",
-    "users_devices",
+    "assets",
     "adapters",
     "enforcements",
     "mixins",
+    "system",
 )